requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
idna==2.10
python-telegram-bot==13.3
pytz==2021.1
requests==2.25.1
six==1.15.0
soupsieve==2.2
telegraph==1.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.3
